+++ tools/emile-map-set.c
@@ -5,6 +5,7 @@
  */
 
 #include <stdlib.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <sys/types.h>
 #include <sys/stat.h>
@@ -252,7 +253,7 @@ static int put_driver(map_t *map, int partition, char* appledriver)
 	code = malloc(st.st_size);
 	if (code == NULL)
 	{
-		fprintf(stderr, "ERROR: cannot malloc %jd\n", st.st_size);
+		fprintf(stderr, "ERROR: cannot malloc %jd\n", (uintmax_t)st.st_size);
 		return -1;
 	}
